Switch Double to Int on pricing properties
Still figuring out if it's the best way

+++ kotlin/project/modules/domain/src/main/kotlin/payment/Charge.kt
@@ -9,7 +9,8 @@ data class Charge(
     val memberId: Long? = null,
     // se for sponsor, o id do sponsor fica aqui
     val sponsorshipId: Long? = null,
-    val value: Double,
+    // Valor em cêntimos (ex: 150 = 1.50€)
+    val value: Int,
     val status: ChargeStatus,
     // melhor do que ter data: LocalDate, assim sabemos a mensalidade do mes que se ta a pagar e o mes da quota
     // "2025/2026"
