+++ kotlin/project/modules/domain/src/main/kotlin/payment/Payment.kt
@@ -5,7 +5,8 @@ import kotlinx.datetime.LocalDateTime
 data class Payment(
     val paymentId: Long,
     val chargeId: Long,
-    val amount: Double,
+    // Montante em cêntimos (ex: 150 = 1.50€)
+    val amount: Int,
     // STRIPE é a plataforma que faz as transações do dinheiro, campo só dedicado a ele
     val provider: String,
     // sessionId/intentId
